requirements.txt
torch==1.8.0+cu111
torchvision==0.9.0+cu111
torchaudio==0.8.0


numpy==1.26.4
pandas==2.3.1
scipy==1.13.1
scikit-learn==1.6.1


h5py==3.14.0
ujson==5.11.0


matplotlib==3.5.1
seaborn==0.11.2


cvxpy==1.7.2


transformers==4.57.1
datasets==4.4.1


hdbscan==0.8.29


calmsize==0.1.3
click==8.1.8
tqdm==4.67.1
Pillow==11.3.0
PyYAML==6.0.2
requests==2.32.5

